@@ -1,13 +1,8 @@
-#!/usr/bin/env python3
 # --------------------------------------------------------------------------------------
 # SPDX-FileCopyrightText: 2021 Magenta ApS <https://magenta.dk>
 # SPDX-License-Identifier: MPL-2.0
 # --------------------------------------------------------------------------------------
-from ra_flatfile_importer.models import MOFlatFileFormat
-from ra_flatfile_importer.models import MOFlatFileFormatChunk
+from ra_flatfile_importer.cli import cli
 
-
-__all__ = [
-    "MOFlatFileFormat",
-    "MOFlatFileFormatChunk",
-]
+if __name__ == "__main__":
+    cli()
